server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
werkzeug==2.3.7
google-generativeai>=0.3.0
elevenlabs>=0.3.0
SpeechRecognition>=3.10.0
openai>=1.12.0
